feat: add support for Yandex Object Storage mounts with new input syntax
