+++ __tests__/memory.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from '@jest/globals'
-import { GB, MB, parseMemory } from '../src/memory'
+import { GB, MB, parseMemory } from '../src/parse'
 
 const mbs = ['mb', 'MB', ' mb']
 const gbs = ['GB', 'Gb', ' gb']
@@ -8,7 +8,7 @@ test.each(
     [128, 256, 512, 1024, 2048].flatMap(x =>
         mbs.map(u => ({
             input: x.toString() + u,
-            expected: (x as any) * MB
+            expected: x * MB
         }))
     )
 )('memory value $input', ({ input, expected }) => {
@@ -19,7 +19,7 @@ test.each(
     [1, 2, 4, 8].flatMap(x =>
         gbs.map(u => ({
             input: x.toString() + u,
-            expected: (x as any) * GB
+            expected: x * GB
         }))
     )
 )('memory value $input', ({ input, expected }) => {
